Move footer_script_snippet helper function
