+++ ckanext/opendata_theme/opengov_custom_footer/plugin/__init__.py
@@ -1,6 +1,5 @@
 import ckan.plugins as plugins
 import ckan.plugins.toolkit as toolkit
-import re
 
 import ckanext.opendata_theme.base.helpers as helper
 from ckanext.opendata_theme.opengov_custom_footer.controller import CustomFooterController
@@ -49,7 +48,6 @@ class OpenDataThemeFooterPlugin(MixinPlugin):
     def get_helpers(self):
         return {
             'opendata_theme_get_footer_data': get_footer_data,
-            'opendata_theme_get_footer_script_snippet': get_footer_script_snippet,
             'version': helper.version_builder,
         }
 
@@ -59,17 +57,6 @@ def get_footer_data(section):
     if data_dict.get(section):
         return literal(data_dict.get(section))
     return ''
-
-
-def get_footer_script_snippet():
-    pattern = r'<script\b[^>]*>(.*?)<\/script>'
-    script_snippet = toolkit.config.get('ckanext.opendata_theme.script_snippet', '')
-    if not script_snippet:
-        return False
-    match = re.match(pattern, script_snippet, re.IGNORECASE)
-    if not bool(match):
-        return False
-    return literal(script_snippet)
 
 
 def custom_footer_validator(value):
